pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "minirag"
version = "0.2.0"
description = "Explore how your notes connect to each other and surface real-time clusters from your Obsidian vault."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "nicegui",
    "openai",
    "chromadb",
    "python-frontmatter",
    "python-dotenv",
    "umap-learn",
    "hdbscan",
    "numpy",
    "langchain-openai",
    "langchain-core",
]

[project.scripts]
minirag = "app:serve"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["venv*"]
